defined-8/fix: правки замечаний на PR
