+++ src/pages/Forum/Forum.scss
@@ -1,9 +1,9 @@
 .forum {
-  height: 100%;
+  min-height: 100%;
   display: flex;
   flex-direction: column;
   align-items: center;
-  padding-top: 80px;
+  padding: 80px 0 200px;
   box-sizing: border-box;
   background: linear-gradient(to right, #ece9e6, #fff);
