+++ src/pages/Forum/components/ForumTopic/ForumTopic.scss
@@ -1,6 +1,6 @@
 .forum-topic {
   width: 100%;
-  height: 100%;
+  min-height: 100%;
   padding: 0 10px;
   box-sizing: border-box;
 
@@ -31,6 +31,7 @@
     border: none;
     background-color: #f0f4f7;
     border-radius: 10px;
+    align-self: flex-end;
 
     &:hover {
       cursor: pointer;
@@ -39,8 +40,9 @@
   }
 
   &__textarea {
-    margin-left: 5px;
     width: 90%;
     border: none;
+    height: 100px;
+    margin-right: 10px;
   }
 }
