+++ src/pages/Forum/components/ForumTopicCard/ForumTopicCard.scss
