+++ src/pages/Forum/components/ForumTopics/ForumTopics.scss
@@ -21,11 +21,11 @@
   }
 
   &__input {
-    margin-left: 10px;
+    width: 30%;
   }
 
   &__textarea {
-    margin-left: 10px;
-    width: 100%;
+    margin: 0 10px;
+    width: 70%;
   }
 }
